backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
celery==5.3.4

# HTTP clients
httpx==0.25.2
requests==2.31.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Telegram Bot
python-telegram-bot==20.7

# AI and ML
ollama==0.1.7

# Image processing and OCR
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0